[feat] 상점 구매 여부 저장 uint 자료형으로 변경
이후 확장 필요 시 ulong 배열을 통해 구현 예정

+++ Assets/00_Scripts/Manager/SubManager/ShopManager.cs
@@ -187,7 +187,7 @@ public class ShopManager : MonoBehaviour
         PlayerSaveData saveData = SaveSystem.LoadPlayerData();
 
         Log.LogMessage($"{shopItemId}, {saveData.purchaseList}");
-        saveData.purchaseList |= 1 << shopItemId;
+        saveData.purchaseList |= (uint)(1 << shopItemId);
 
         SetMoneyText(saveData.money);
